requirements.txt
asgiref==3.9.1
Django==4.2.24
djangorestframework==3.16.1
sqlparse==0.5.3
typing_extensions==4.15.0
django-cors-headers==4.7.0
httpx==0.27.2
python-dotenv==1.1.1